Update auth modal
