+++ frontend/src/lib/helpers.js
@@ -6,3 +6,9 @@ export function formatDate(dateString) {
 		day: 'numeric'
 	});
 }
+
+export function preventNonNumericInput(e) {
+	if (!/[0-9]/.test(e.key)) {
+		e.preventDefault();
+	}
+}
